Move teleterm integration tests to integration/proxy
