+++ integration/proxy/proxy_test.go
@@ -814,6 +814,10 @@ func TestALPNSNIProxyDatabaseAccess(t *testing.T) {
 		// Disconnect.
 		require.NoError(t, client.Close())
 	})
+
+	t.Run("teleterm gateways cert renewal", func(t *testing.T) {
+		testTeletermGatewaysCertRenewal(t, pack)
+	})
 }
 
 // TestALPNSNIProxyAppAccess tests application access via ALPN SNI proxy service.
